reveive control robot msg and waypoint msg from thingsboard, foward to UAV

+++ communication/protocols/messages/uavlink_msg_setwp_t.h
@@ -10,7 +10,7 @@ class uavlink_msg_setwp_t : public QObject
 {
     Q_OBJECT
 public:
-    explicit uavlink_msg_setwp_t(qint64 type, qint64 wpId, float targetX, float targetY,float targetZ, QObject *parent = nullptr);
+    explicit uavlink_msg_setwp_t(qint64 type, qint64 wpId, qint64 targetX, qint64 targetY,qint64 targetZ, QObject *parent = nullptr);
     ~uavlink_msg_setwp_t();
     QByteArray ToPackage();
 
@@ -18,9 +18,9 @@ signals:
 private:
     uint16_t _type;
     uint32_t _wpId;
-    float _targetX;
-    float _targetY;
-    float _targetZ;
+    int32_t _targetX;
+    int32_t _targetY;
+    int32_t _targetZ;
 };
 
 #endif // UAVLINK_MSG_SETWP_T_H
